console,account: Move collaborator form
