+++ pkg/webui/console/views/application-collaborator-edit/application-collaborator-edit.js
@@ -19,7 +19,8 @@ import { useBreadcrumbs } from '@ttn-lw/components/breadcrumbs/context'
 import PageTitle from '@ttn-lw/components/page-title'
 import Breadcrumb from '@ttn-lw/components/breadcrumbs/breadcrumb'
 import toast from '@ttn-lw/components/toast'
-import CollaboratorForm from '@ttn-lw/components/collaborator-form'
+
+import CollaboratorForm from '@ttn-lw/containers/collaborator-form'
 
 import sharedMessages from '@ttn-lw/lib/shared-messages'
 import PropTypes from '@ttn-lw/lib/prop-types'
