+++ pkg/webui/console/views/gateway-collaborator-edit/gateway-collaborator-edit.js
@@ -19,7 +19,8 @@ import PageTitle from '@ttn-lw/components/page-title'
 import Breadcrumb from '@ttn-lw/components/breadcrumbs/breadcrumb'
 import toast from '@ttn-lw/components/toast'
 import { useBreadcrumbs } from '@ttn-lw/components/breadcrumbs/context'
-import CollaboratorForm from '@ttn-lw/components/collaborator-form'
+
+import CollaboratorForm from '@ttn-lw/containers/collaborator-form'
 
 import sharedMessages from '@ttn-lw/lib/shared-messages'
 import PropTypes from '@ttn-lw/lib/prop-types'
